refactor: 재고 관리와 예약 로직을 서비스 단위로 분리
