+++ catalog-service/src/main/java/com/node5/catalogservice/inventory/presentation/InventoryInternalController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.node5.catalogservice.inventory.application.InventoryService;
+import com.node5.catalogservice.inventory.application.InventoryReservationService;
 import com.node5.catalogservice.inventory.application.dto.StockHoldBatchResult;
 import com.node5.catalogservice.inventory.presentation.dto.StockCommitRequest;
 import com.node5.catalogservice.inventory.presentation.dto.StockHoldBatchRequest;
@@ -26,7 +26,7 @@ import lombok.RequiredArgsConstructor;
 @Tag(name = "Inventory Internal", description = "주문/결제 흐름을 위한 내부 재고 예약 처리 API")
 public class InventoryInternalController {
 
-	private final InventoryService inventoryService;
+	private final InventoryReservationService inventoryReservationService;
 
 	@PostMapping("/hold")
 	@Operation(
@@ -39,7 +39,7 @@ public class InventoryInternalController {
 		@ApiResponse(responseCode = "404", description = "상품 재고가 존재하지 않습니다.")
 	})
 	public ResponseEntity<StockHoldBatchResult> hold(@Valid @RequestBody StockHoldBatchRequest request) {
-		var result = inventoryService.holdBatch(request.toCommand());
+		var result = inventoryReservationService.holdBatch(request.toCommand());
 		return ResponseEntity.status(HttpStatus.CREATED).body(result);
 	}
 
@@ -54,7 +54,7 @@ public class InventoryInternalController {
 		@ApiResponse(responseCode = "404", description = "재고 예약이 존재하지 않습니다.")
 	})
 	public ResponseEntity<Void> commit(@Valid @RequestBody StockCommitRequest request) {
-		inventoryService.commit(request.toCommand());
+		inventoryReservationService.commit(request.toCommand());
 		return ResponseEntity.ok().build();
 	}
 
@@ -69,7 +69,7 @@ public class InventoryInternalController {
 		@ApiResponse(responseCode = "404", description = "재고 예약이 존재하지 않습니다.")
 	})
 	public ResponseEntity<Void> release(@Valid @RequestBody StockReleaseRequest request) {
-		inventoryService.release(request.toCommand());
+		inventoryReservationService.release(request.toCommand());
 		return ResponseEntity.ok().build();
 	}
 }
